feat: add basic socket generation from cf
