+++ bin/si-gen-cloud-control/src/errors.ts
@@ -4,3 +4,12 @@ export class CommandFailed extends Error {
     this.name = "CommandFailed";
   }
 }
+
+export class ServiceMissing extends Error {
+  constructor(serviceName: string) {
+    super(
+      `Attempt to find schema for service ${serviceName}, but it does not exist`,
+    );
+    this.name = "SchemaMissing";
+  }
+}
